+++ templemeads/src/agent_bridge.rs
@@ -39,7 +39,7 @@ pub async fn run(config: Config, runner: AsyncRunnable) -> Result<(), Error> {
     }
 
     // pass the service details onto the handler
-    set_my_service_details(&config.service.name(), &config.agent, Some(runner)).await?;
+    set_my_service_details(&config.service.name(), &config.agent, Some(runner), true).await?;
 
     // spawn the bridge server
     spawn(config.bridge).await?;