+++ templemeads/src/portal.rs
@@ -24,7 +24,13 @@ pub async fn run(config: Config, runner: AsyncRunnable) -> Result<(), Error> {
     }
 
     // pass the service details onto the handler
-    set_my_service_details(&config.service().name(), &config.agent(), Some(runner)).await?;
+    set_my_service_details(
+        &config.service().name(),
+        &config.agent(),
+        Some(runner),
+        true,
+    )
+    .await?;
 
     // run the Portal OpenPortal agent
     paddington::set_handler(process_message).await?;